+++ core/src/main/java/com/hujiugame/qingfeng/core/GameHost.java
@@ -103,7 +103,6 @@ public final class GameHost
                 this.textManager,
                 this.audioManager,
                 this.graphicsManager,
-                this.uiManager,
                 this.layoutManager,
                 this.playLocalData),
             new GamePlayDataLoader(