+++ core/src/main/java/com/hujiugame/qingfeng/scene/impl/ConfigBasic.java
@@ -9,7 +9,7 @@ import com.hujiugame.qingfeng.input.VirtualInputHandler;
 import com.hujiugame.qingfeng.scene.GameRender;
 import com.hujiugame.qingfeng.ui.UiManager;
 import com.hujiugame.qingfeng.event.EventQueue;
-import com.hujiugame.qingfeng.event.imp.EventPopGameState;
+import com.hujiugame.qingfeng.event.imp.PopGameState;
 import com.hujiugame.qingfeng.util.system.LogUtils;
 
 import java.util.ArrayList;
@@ -116,7 +116,7 @@ public final class ConfigBasic implements GameRender
         // 按下返回按钮
         if (uiManager.isButtonClicked(UniversalUiKey.BUTTON_BACK))
         {
-            eventQueue.addEvent(new EventPopGameState());
+            eventQueue.addEvent(new PopGameState());
         }
     }